Add support for both MUSL and GLIBC with two distro options: adi-distro-musl and adi-distro-glibc
Note: GLIBC builds are currently too large to fit into SPI flash partition allocations on ezkits,
so you must boot via TFTP, eMMC, SD, USB, or some other mechanism (custom board)

+++ meta-adi-adsp-sc5xx/classes/adsp-sc5xx.bbclass
@@ -56,8 +56,10 @@ IMAGE_FSTYPES:append = " tar.xz jffs2 "
 #For some reason on poky-tiny, INIT_MANAGER="systemd" does not appear to work
 #For now, let's relink directly to systemd
 fakeroot do_set_init(){
-    rm -rf ${IMAGE_ROOTFS}/sbin/init
-    ln -s /lib/systemd/systemd ${IMAGE_ROOTFS}/sbin/init
+    if [ "${DISTRO}" = "adi-distro-musl" ]; then
+        rm -rf ${IMAGE_ROOTFS}/sbin/init
+        ln -s /lib/systemd/systemd ${IMAGE_ROOTFS}/sbin/init
+    fi
 }
 
 addtask do_set_init after do_rootfs before do_image
